+++ package.json
@@ -10,10 +10,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "^19.2.4",
-    "react-dom": "^19.2.4",
-    "react-router-dom": "^7.13.1",
-    "styled-components": "^6.3.11"
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "react-router-dom": "^7.13.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.39.1",
@@ -24,11 +23,7 @@
     "eslint-plugin-react-hooks": "^7.0.1",
     "eslint-plugin-react-refresh": "^0.4.24",
     "globals": "^16.5.0",
-    "prettier": "^3.8.1",
-    "vite": "^8.0.0-beta.13",
+    "vite": "^7.3.1",
     "vite-plugin-svgr": "^4.5.0"
-  },
-  "overrides": {
-    "vite": "^8.0.0-beta.13"
   }
 }
