+++ src/App.jsx
@@ -1,14 +1,12 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Header from "./components/Layout/Header";
 import HomePage from "./pages/HomePage/HomePage";
 import LoginPage from "./pages/LoginPage/LoginPage";
 import MarketPage from "./pages/MarketPage/MarketPage";
-import AdditemPage from "./pages/AdditemPage/AdditemPage";
 import CommunityPage from "./pages/CommunityPage/CommunityPage";
-import Header from "./components/Layout/Header.jsx";
-
+import AdditemPage from "./pages/AdditemPage/AdditemPage";
 
 function App() {
-  
   return (
     <BrowserRouter>
       <Header />
@@ -23,7 +21,7 @@ function App() {
         </Routes>
       </div>
     </BrowserRouter>
-  )
+  );
 }
 
 export default App;
