+++ src/api/itemApi.js
@@ -1,19 +1,17 @@
 export async function getProducts(params = {}) {
+  const query = new URLSearchParams(params).toString();
 
-    const query = new URLSearchParams(params).toString();
-
-    try {
-        const res = await fetch(
-            `https://panda-market-api.vercel.app/products?${query}`
-        );
-        if(!res.ok) {
-            throw new Error(`HTTP ERROR: ${res.status}`);
-        }
-
-        const data = await res.json();
-        return data;
-    } catch (error) {
-        console.error(`fetch error`, error);
-        throw error;
+  try {
+    const response = await fetch(
+      `https://panda-market-api.vercel.app/products?${query}`,
+    );
+    if (!response.ok) {
+      throw new Error(`HTTP error: ${response.status}`);
     }
+    const body = await response.json();
+    return body;
+  } catch (error) {
+    console.error("Failed to fetch products:", error);
+    throw error;
+  }
 }
