+++ src/components/Layout/Header.css
@@ -1,55 +1,54 @@
 .headerLeft {
-    display: flex;
-    align-items: center;
+  display: flex;
+  align-items: center;
 }
 
 .headerLogo {
-    margin-right: 16px;
+  margin-right: 16px;
 }
 
 .globalHeader nav ul {
-    display: flex;
-    list-style: none;
-    gap: 8px;
-    font-weight: bold;
-    font-size: 16px;
-    color: #4b5563;
+  display: flex;
+  list-style: none;
+  gap: 8px;
+  font-weight: bold;
+  font-size: 16px;
+  color: #4b5563;
 }
 
 .globalHeader nav ul li a:hover {
-    color: var(--blue);
+  color: var(--blue);
 }
 
 .loginLink,
 .registerButton {
-    font-size: 16px;
-    font-weight: 600;
-    border-radius: 8px;
-    padding: 11.5px 23px;
+  font-size: 16px;
+  font-weight: 600;
+  border-radius: 8px;
+  padding: 11.5px 23px;
 }
 
 @media (min-width: 768px) {
-    .globalHeader nav ul {
-        gap: 36px;
-        font-size: 18px;
-    }
+  .globalHeader nav ul {
+    gap: 36px;
+    font-size: 18px;
+  }
 
-    .headerLogo {
-        margin-right: 35px;
-    }
+  .headerLogo {
+    margin-right: 35px;
+  }
 
-    .headerLogo img {
-        width: 153px;
-    }
-
+  .headerLogo img {
+    width: 153px;
+  }
 }
 
 @media (min-width: 1280px) {
-    .headerLogo {
-        margin-right: 47px;
-    }
+  .headerLogo {
+    margin-right: 47px;
+  }
 
-    .headerLogo img {
-        width: 153px;
-    }
+  .headerLogo img {
+    width: 153px;
+  }
 }
