--- src/components/MarketPageUI/DropdownList.css
@@ -1,22 +1,0 @@
-.dropdownList {
-    position: absolute;
-    top: 110%;
-    right: 0;
-    background: #fff;
-    border-radius: 8px;
-    border: 1px solid #e5e7eb;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    z-index: 99;
-}
-
-.dropdownItem {
-    padding: 12px 44px;
-    border-bottom: 1px solid #e5e7eb;
-    font-size: 16px;
-    color: #1f2937;
-    cursor: pointer;
-}
-
-.dropdownItem:last-child {
-    border-bottom: none;
-}
