--- src/components/MarketPageUI/PaginationBar.jsx
@@ -1,52 +1,0 @@
-import './PaginationBar.css';
-import LeftArrow from '../../assets/images/marketIcons/arrow_left.svg?react';
-import RightArrow from '../../assets/images/marketIcons/arrow_right.svg?react';
-
-const PaginationBar = ({ totalPageNum, activePageNum, onPageChange }) => {
-    const maxVisiblePages = 5;
-    let startPage;
-
-    if (totalPageNum <= maxVisiblePages) {
-        startPage = 1;
-    } else {
-        startPage = Math.max(activePageNum - Math.floor(maxVisiblePages / 2), 1);
-        startPage = Math.min(startPage, totalPageNum - maxVisiblePages + 1);
-    }
-
-    const pages = Array.from(
-        { length: Math.min(maxVisiblePages, totalPageNum - startPage + 1) },
-        (_, i) => startPage + i
-    );
-
-    return (
-        <div className="paginationBar">
-            <button
-                className="paginationButton"
-                disabled={activePageNum === 1}
-                onClick={() => onPageChange(activePageNum - 1)}
-            >
-                <LeftArrow />
-            </button>
-            {pages.map((page) => (
-                <button
-                    key={page}
-                    className={`paginationButton ${
-                        activePageNum === page ? 'active' : ''
-                    }`}
-                    onClick={() => onPageChange(page)}
-                >
-                    {page}
-                </button>
-            ))}
-            <button
-                className="paginationButton"
-                disabled={activePageNum === totalPageNum}
-                onClick={() => onPageChange(activePageNum + 1)}
-            >
-                <RightArrow />
-            </button>
-        </div>
-    )
-}
-
-export default PaginationBar;
