+++ src/components/MarketUI/DropdownList.jsx
@@ -1,0 +1,14 @@
+import "./DropdownList.css";
+
+export default function DropdownList({ onSortSelection }) {
+  return (
+    <div className="dropdownList">
+      <div className="dropdownItem" onClick={() => onSortSelection('recent')}>
+        최신순
+      </div>
+      <div className="dropdownItem" onClick={() => onSortSelection('favorite')}>
+        인기순
+      </div>
+    </div>
+  )
+}
