+++ src/pages/AdditemPage/AdditemPage.jsx
@@ -1,7 +1,3 @@
-import './AdditemPage.css';
-
-function AdditemPage() {
-    return <div>AdditemPage</div>
+export default function AdditemPage() {
+  return <div>AdditemPage</div>
 }
-
-export default AdditemPage;
