+++ src/pages/CommunityPage/CommunityPage.jsx
@@ -1,5 +1,3 @@
-function CommunityPage() {
-    return <div>CommunityPage</div>
+export default function CommunityPage() {
+  return <div>CommunityPage</div>
 }
-
-export default CommunityPage;
