+++ src/pages/HomePage/HomePage.jsx
@@ -1,7 +1,3 @@
-import './HomePage.css';
-
-function HomePage() {
-    return <div>HomePage</div>
+export default function HomePage() {
+  return <div>HomePage</div>
 }
-
-export default HomePage;
