+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,7 +1,3 @@
-import './LoginPage.css';
-
-function LoginPage() {
-    return <div>LoginPage</div>
+export default function LoginPage() {
+  return <div>LoginPage</div>
 }
-
-export default LoginPage;
