+++ src/pages/MarketPage/MarketPage.css
@@ -1,190 +1,190 @@
 .sectionTitle {
-    color: #111827;
-    font-weight: bold;
-    font-size: 20px;
+  color: #111827;
+  font-weight: bold;
+  font-size: 20px;
 }
 
 .itemCard {
-    color: #1f2937;
-    overflow: hidden;
-    cursor: pointer;
+  color: #1f2937;
+  overflow: hidden;
+  cursor: pointer;
 }
 
 .itemCardThumbnail {
-    width: 100%;
-    height: auto;
-    object-fit: cover;
-    border-radius: 16px;
-    overflow: hidden;
-    aspect-ratio: 1;
-    margin-bottom: 16px;
+  width: 100%;
+  height: auto;
+  object-fit: cover;
+  border-radius: 16px;
+  overflow: hidden;
+  aspect-ratio: 1;
+  margin-bottom: 16px;
 }
 
 .itemSummary {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  flex-grow: 1;
 }
 
 .itemName {
-    font-size: 16px;
-    font-weight: 400;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
+  font-size: 16px;
+  font-weight: 400;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
 }
 
 .itemPrice {
-    font-size: 16px;
-    font-weight: bold;
+  font-size: 16px;
+  font-weight: bold;
 }
 
 .favoriteCount {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    color: #4b5563;
-    font-size: 12px;
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  color: #4b5563;
+  font-size: 12px;
 }
 
 .bestItemsContainer {
-    padding-top: 17px;
-    padding-bottom: 24px;
+  padding-top: 17px;
+  padding-bottom: 24px;
 }
 
 .allItemsCardSection {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 32px 8px;
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 32px 8px;
 }
 
 .allItemsHeaderWrapper {
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    margin-bottom: 16px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-bottom: 16px;
 }
 
 .topRow,
 .bottomRow {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 }
 
 .sortButtonWrapper {
-    position: relative;
+  position: relative;
 }
 
 .sortDropdownTriggerButton {
-    border: 1px solid #e5e7eb;
-    border-radius: 12px;
-    padding: 9px;
-    margin-left: 8px;
+  border: 1px solid #e5e7eb;
+  border-radius: 12px;
+  padding: 9px;
+  margin-left: 8px;
 }
 
 .searchBarWrapper {
-    display: flex;
-    background-color: #f3f4f6;
-    border-radius: 12px;
-    padding: 9px 16px;
-    flex: 1;
-    align-items: center;
-    margin-right: 8px;
+  display: flex;
+  background-color: #f3f4f6;
+  border-radius: 12px;
+  padding: 9px 16px;
+  flex: 1;
+  align-items: center;
+  margin-right: 8px;
 }
 
 .searchBarInput {
-    border: none;
-    flex: 1;
-    background-color: inherit;
-    margin-left: 4px;
+  border: none;
+  flex: 1;
+  background-color: inherit;
+  margin-left: 4px;
 }
 
 .searchBarInput::placeholder {
-    color: #9ca3af;
-    font-size: 16px;
+  color: #9ca3af;
+  font-size: 16px;
 }
 
 .searchBarInput:focus {
-    outline: none;
+  outline: none;
 }
 
 .paginationBarWrapper {
-    padding-top: 40px;
-    padding-bottom: 80px;
+  padding-top: 40px;
+  padding-bottom: 80px;
 }
 
 @media (min-width: 768px) {
-   .allItemsHeaderWrapper {
-        position: relative;
-        flex-direction: row;
-        align-items: center;
-    }
+  .allItemsHeaderWrapper {
+    position: relative;
+    flex-direction: row;
+    align-items: center;
+  }
 
-    .bottomRow {
-        margin-left: auto;
-        gap: 148px;
-    }
+  .bottomRow {
+    margin-left: auto;
+    gap: 148px;
+  }
 
-    .registerButton {
-        position: absolute;
-        right: 60px;  /* 정렬 버튼 왼쪽에 위치 */
-        top: 50%;
-        transform: translateY(-50%);
-    }
+  .registerButton {
+    position: absolute;
+    right: 60px; /* 정렬 버튼 왼쪽에 위치 */
+    top: 50%;
+    transform: translateY(-50%);
+  }
 
-    .searchBarWrapper {
-        width: 242px;
-        flex: none;
-    }
+  .searchBarWrapper {
+    width: 242px;
+    flex: none;
+  }
 
-    .topRow {
-        gap: 20px;
-    }
+  .topRow {
+    gap: 20px;
+  }
 
-    .bestItemsContainer {
-        margin-bottom: 40px;
-    }
+  .bestItemsContainer {
+    margin-bottom: 40px;
+  }
 
-    .bestItemsCardSection {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        gap: 24px;
-    }
+  .bestItemsCardSection {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 24px;
+  }
 
-    .allItemsCardSection {
-        grid-template-columns: repeat(3, 1fr);
-        gap: 40px 16px;
-    }
+  .allItemsCardSection {
+    grid-template-columns: repeat(3, 1fr);
+    gap: 40px 16px;
+  }
 }
 
 @media (min-width: 1280px) {
-    .allItemsHeaderWrapper {
-        flex-direction: row;
-        align-items: center;
-    }
+  .allItemsHeaderWrapper {
+    flex-direction: row;
+    align-items: center;
+  }
 
-    .bottomRow {
-        margin-left: auto;
-        gap: 148px;
-    }
+  .bottomRow {
+    margin-left: auto;
+    gap: 148px;
+  }
 
-    .searchBarWrapper {
-        width: 325px;
-        flex: none;
-    }
+  .searchBarWrapper {
+    width: 325px;
+    flex: none;
+  }
 
-    .topRow {
-        gap: 20px;
-    }
+  .topRow {
+    gap: 20px;
+  }
 
-    .bestItemsCardSection {
-        grid-template-columns: repeat(4, 1fr);
-    }
+  .bestItemsCardSection {
+    grid-template-columns: repeat(4, 1fr);
+  }
 
-    .allItemsCardSection {
-        grid-template-columns: repeat(5, 1fr);
-        gap: 40px 24px;
-    }
+  .allItemsCardSection {
+    grid-template-columns: repeat(5, 1fr);
+    gap: 40px 24px;
+  }
 }
