+++ src/pages/MarketPage/components/ItemCard.jsx
@@ -1,19 +1,19 @@
-import HeartIcon from '../../../assets/images/marketIcons/ic_heart.svg?react';
+import HeartIcon from "../../../assets/icons/ic_heart.svg?react";
 
 function ItemCard({ item }) {
-    return (
-        <div className="itemCard">
-            <img src={item.images[0]} alt={item.name} className="itemCardThumbnail" />
-            <div className='itemSummary'>
-                <h2 className='itemName'>{item.name}</h2>
-                <p className='itemPrice'>{item.price.toLocaleString()}원</p>
-                <div className='favoriteCount'>
-                    <HeartIcon />
-                    {item.favoriteCount}
-                </div>
-            </div>
+  return (
+    <div className="itemCard">
+      <img src={item.images[0]} alt={item.name} className="itemCardThumbnail" />
+      <div className="itemSummary">
+        <h2 className="itemName">{item.name}</h2>
+        <p className="itemPrice">{item.price.toLocaleString()}원</p>
+        <div className="favoriteCount">
+          <HeartIcon />
+          {item.favoriteCount}
         </div>
-    )
+      </div>
+    </div>
+  );
 }
 
 export default ItemCard;
