+++ vite.config.js
@@ -6,6 +6,6 @@ import svgr from 'vite-plugin-svgr'
 export default defineConfig({
   plugins: [
     react(),
-    svgr(),
+    svgr()
   ],
 })
